rag-chatbot/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
qdrant-client
psycopg2-binary
langchain
tiktoken
